refactor: cardを共通化
