--- src/components/organisms/card/RouteCard/style.module.css
@@ -1,41 +1,0 @@
-.container {
-  margin: 20px 0;
-  display: flex;
-  padding: 10px;
-  border: solid;
-  border-radius: 5px;
-  border-width: 1px;
-  cursor: pointer;
-}
-
-.image-container {
-  width: 200px;
-}
-
-.info-container {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-}
-
-.header {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: flex-end;
-}
-
-.name {
-  font-size: 2.5rem;
-  margin-top: 0;
-  margin-bottom: 0;
-}
-
-.update {
-  margin-right: 10px;
-  margin-bottom: 0;
-}
-
-.info {
-  margin: 5px 0;
-}
