+++ src/components/organisms/cardList/RouteCardList/index.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { searchRoutes } from "../../../../api/routes";
 import { RouteInfo } from "../../../../types";
-import RouteCard from "../RouteCard";
+import RouteCard from "../../RouteCard";
 import styles from "./style.module.css";
 
 type EouteCardListProps = {
